Move PrettyM and ScopeBoundary to new module
