+++ src/Covenant/Internal/Type.hs
@@ -12,7 +12,6 @@ module Covenant.Internal.Type
     ValT (..),
     BuiltinFlatT (..),
     TyName (..),
-    ScopeBoundary (..), -- used in the generators
     DataEncoding (..),
     runConstructorName,
     abstraction,
@@ -26,12 +25,7 @@ module Covenant.Internal.Type
   )
 where
 
-import Control.Monad.Reader
-  ( MonadReader (local),
-    Reader,
-    asks,
-    runReader,
-  )
+import Control.Monad.Reader (asks, local)
 import Covenant.DeBruijn (DeBruijn (Z))
 import Covenant.Index
   ( Count,
@@ -40,6 +34,11 @@ import Covenant.Index
     intCount,
     intIndex,
     ix0,
+  )
+import Covenant.Internal.PrettyPrint
+  ( PrettyM,
+    ScopeBoundary (ScopeBoundary),
+    runPrettyM,
   )
 import Covenant.Internal.Strategy
   ( DataEncoding
@@ -64,7 +63,6 @@ import Covenant.Internal.Strategy
 import Covenant.Util (pattern ConsV, pattern NilV)
 import Data.Functor.Classes (Eq1 (liftEq))
 import Data.Kind (Type)
-import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Set (Set)
 import Data.String (IsString)
@@ -447,69 +445,6 @@ byteStringBaseFunctor = DataDeclaration "ByteString_F" count1 constrs SOP
       ]
 
 -- Helpers
-
-newtype ScopeBoundary = ScopeBoundary Int
-  deriving (Show, Eq, Ord, Num, Real, Enum, Integral) via Int
-
--- Keeping the field names for clarity even if we don't use them
-data PrettyContext (ann :: Type)
-  = PrettyContext
-  { _boundIdents :: Map ScopeBoundary (Vector (Doc ann)),
-    _currentScope :: ScopeBoundary,
-    _varStream :: [Doc ann]
-  }
-
-instance
-  (k ~ A_Lens, a ~ Map ScopeBoundary (Vector (Doc ann)), b ~ Map ScopeBoundary (Vector (Doc ann))) =>
-  LabelOptic "boundIdents" k (PrettyContext ann) (PrettyContext ann) a b
-  where
-  {-# INLINEABLE labelOptic #-}
-  labelOptic =
-    lens
-      (\(PrettyContext x _ _) -> x)
-      (\(PrettyContext _ y z) x -> PrettyContext x y z)
-
-instance
-  (k ~ A_Lens, a ~ ScopeBoundary, b ~ ScopeBoundary) =>
-  LabelOptic "currentScope" k (PrettyContext ann) (PrettyContext ann) a b
-  where
-  {-# INLINEABLE labelOptic #-}
-  labelOptic =
-    lens
-      (\(PrettyContext _ x _) -> x)
-      (\(PrettyContext x _ z) y -> PrettyContext x y z)
-
-instance
-  (k ~ A_Lens, a ~ [Doc ann], b ~ [Doc ann]) =>
-  LabelOptic "varStream" k (PrettyContext ann) (PrettyContext ann) a b
-  where
-  {-# INLINEABLE labelOptic #-}
-  labelOptic =
-    lens
-      (\(PrettyContext _ _ x) -> x)
-      (\(PrettyContext x y _) z -> PrettyContext x y z)
-
--- Maybe make a newtype with error reporting since this can fail, but do later since *should't* fail
-newtype PrettyM (ann :: Type) (a :: Type) = PrettyM (Reader (PrettyContext ann) a)
-  deriving
-    ( Functor,
-      Applicative,
-      Monad,
-      MonadReader (PrettyContext ann)
-    )
-    via (Reader (PrettyContext ann))
-
-runPrettyM :: forall (ann :: Type) (a :: Type). PrettyM ann a -> a
-runPrettyM (PrettyM ma) = runReader ma (PrettyContext mempty 0 infiniteVars)
-  where
-    -- Lazily generated infinite list of variables. Will start with a, b, c...
-    -- and cycle around to a1, b2, c3 etc.
-    -- We could do something more sophisticated but this should work.
-    infiniteVars :: [Doc ann]
-    infiniteVars =
-      let aToZ = ['a' .. 'z']
-          intStrings = ("" <$ aToZ) <> map (show @Integer) [0 ..]
-       in zipWith (\x xs -> pretty (x : xs)) aToZ intStrings
 
 prettyCompTWithContext :: forall (ann :: Type). CompT Renamed -> PrettyM ann (Doc ann)
 prettyCompTWithContext (CompT count (CompTBody funArgs))
